fix: separate mcp inference handler for auth consistency
