+++ docs/features/governance/virtual-keys.mdx
@@ -488,7 +488,7 @@ curl -X DELETE http://localhost:8080/api/governance/customers/{customer_id}
 
 ## Usage
 
-### Required Header
+### Making Virtual Keys Mandatory
 
 All governance-enabled requests must include the virtual key header:
 
@@ -502,14 +502,16 @@ curl -X POST http://localhost:8080/v1/chat/completions \
   }'
 ```
 
-By default governance is optional, meaning that if the `x-bf-vk` header is not present, the request will be allowed but without any governance checks/routing. But you can make it mandatory by enforcing the governance header.
+By default governance is optional, meaning that if the virtual key header is not present, the request will be allowed but without any governance checks/routing. But you can make it mandatory by enforcing the virtual key header.
 
 <Tabs group="enforce-governance-header">
 <Tab title="Web UI">
 
-1. Go to **Client** → **Governance**
+1. Go to **Config** → **Security**
 
-2. Check the **Enforce Governance Header** checkbox
+2. Check the **Enforce Virtual Keys** checkbox
+
+![Enforce Virtual Keys](../../media/ui-enforce-virtual-keys.png)
 
 </Tab>
 <Tab title="API">
@@ -540,23 +542,82 @@ curl -X PUT http://localhost:8080/api/config \
 
 When the governance header is enforced, the request will be rejected if the `x-bf-vk` header is not present.
 
-### Optional Audit Headers
+### Authentication and Virtual Keys
 
-Include additional headers for enhanced tracking and audit trails:
+Virtual keys and HTTP authentication are **independent layers** that can work together:
+
+| Layer | Purpose | Headers |
+|-------|---------|---------|
+| **Authentication** | Validates user identity | `Authorization: Basic/Bearer <credentials>` |
+| **Virtual Keys** | Request routing and governance | `x-bf-vk`, `Authorization`, `x-api-key`, `x-goog-api-key` |
+
+**When `disable_auth_on_inference: true` (auth disabled):**
+
+Virtual keys can be passed via any supported header without additional authentication:
+
+```bash
+# Using x-bf-vk header
+curl -X POST http://localhost:8080/v1/chat/completions \
+  -H "x-bf-vk: sk-bf-your-key" \
+  -H "Content-Type: application/json" \
+  -d '{"model": "gpt-4o-mini", "messages": [...]}'
+
+# Using Authorization header (OpenAI style)
+curl -X POST http://localhost:8080/v1/chat/completions \
+  -H "Authorization: Bearer sk-bf-your-key" \
+  -H "Content-Type: application/json" \
+  -d '{"model": "gpt-4o-mini", "messages": [...]}'
+```
+
+**When `disable_auth_on_inference: false` (auth enabled):**
+
+You must provide both authentication credentials AND the virtual key. Use `x-bf-vk` for the virtual key since the `Authorization` header is used for authentication:
 
 ```bash
 curl -X POST http://localhost:8080/v1/chat/completions \
+  -H "Authorization: Basic <base64-credentials>" \
+  -H "x-bf-vk: sk-bf-your-key" \
   -H "Content-Type: application/json" \
-  -H "x-bf-vk: vk-engineering-main" \
-  -H "x-bf-user-id: user-alice" \
+  -d '{"model": "gpt-4o-mini", "messages": [...]}'
+```
+
+**Configuring `disable_auth_on_inference`:**
+
+<Tabs group="config-method">
+<Tab title="Web UI">
+
+1. Go to **Settings** → **Security**
+2. Toggle **Disable Auth on Inference** to enable/disable
+
+![Disable Auth on Inference](../../media/ui-disable-auth-on-inference.png)
+
+</Tab>
+<Tab title="API">
+
+```bash
+curl -X PUT http://localhost:8080/api/config \
+  -H "Content-Type: application/json" \
   -d '{
-    "model": "gpt-4o-mini", 
-    "messages": [{"role": "user", "content": "Hello!"}]
+    "auth_config": {
+      "disable_auth_on_inference": true
+    }
   }'
 ```
 
-**Header Definitions:**
-- `x-bf-vk` - **Required** virtual key for access control
+</Tab>
+<Tab title="config.json">
+
+```json
+{
+  "auth_config": {
+    "is_enabled": true,
+    "disable_auth_on_inference": true
+  }
+}
+```
+
+</Tab>
+</Tabs>
 
 ### Error Responses
